+++ hoot-core/src/main/cpp/hoot/core/info/NumericStatistic.h
@@ -53,7 +53,7 @@ public:
   virtual long getInformationCount() const { return 0; }
   virtual long getInformationMin() const { return 0; }
   virtual long getInformationMax() const { return 0; }
-  virtual long getInformationCountDiff() const { return getStat() - getInformationCount(); }
+  virtual long getInformationCountDiff() const { return static_cast<long>(getStat()) - getInformationCount(); }
   virtual double getInformationAverage() const { return 0; }
 };
