+++ hoot-core/src/main/cpp/hoot/core/util/SignalCatcher.cpp
@@ -117,7 +117,7 @@ void SignalCatcher::print_stacktrace(FILE *out, unsigned int max_frames)
   void* addrlist[max_frames+1];
 
   // retrieve current stack addresses
-  int addrlen = backtrace(addrlist, sizeof(addrlist) / sizeof(void*));
+  int addrlen = backtrace(addrlist, static_cast<int>(sizeof(addrlist) / sizeof(void*)));
 
   if (addrlen == 0)
   {
